% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moccsR.R
\name{calcMOCCS2score}
\alias{calcMOCCS2score}
\title{Calculate MOCCS2score}
\usage{
calcMOCCS2score(fasta, k, ignoreLowerCase = TRUE)
}
\arguments{
\item{fasta}{Input fasta file}

\item{k}{Length of k-mer}

\item{ignoreLowerCase}{Logical parameter. If TRUE, the function ignores
lower case character in fasta file.}
}
\value{
data.frame that consists of kmer, auc, MOCCS2score, p-value, and
q-value.
}
\description{
Calculate MOCCS2score
}
\examples{
fastaPath <- system.file("count.fasta", package = "moccsR")
calcMOCCS2score(fastaPath, k = 6, ignoreLowerCase = TRUE)
}
