% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moccsR.R
\name{calcMOCCS2score}
\alias{calcMOCCS2score}
\title{Calculate MOCCS2score}
\usage{
calcMOCCS2score(fasta, k, ignoreLowerCase = T)
}
\arguments{
\item{fasta}{Input fasta file}

\item{k}{Length of k-mer}

\item{ignoreLowerCase}{Logical parameter. If TRUE, the function ignores
lower case character in fasta file.}
}
\description{
Calculate MOCCS2score
}
